@@ -11,24 +11,25 @@
     "spa"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
+  "exports": "./index.mjs",
   "scripts": {
-    "test": "test-runner test/*.js",
+    "test": "test-runner test/*.mjs",
     "cover": "nyc npm test && nyc report --reporter=text-lcov | coveralls"
   },
   "dependencies": {
-    "koa-send": "^5.0.0"
+    "koa-send": "^5.0.1"
   },
   "devDependencies": {
-    "coveralls": "^3.0.9",
-    "lws": "^2.0.8",
-    "lws-static": "^1.1.3",
-    "node-fetch": "^2.6.0",
-    "nyc": "^15.0.0",
-    "test-runner": "^0.8.13"
+    "coveralls": "^3.1.1",
+    "lws": "^4.0.1",
+    "lws-static": "^3.0.0",
+    "node-fetch": "^2.6.1",
+    "nyc": "^15.1.0",
+    "test-runner": "^0.9.6"
   },
   "files": [
-    "index.js"
+    "index.mjs"
   ]
 }
